requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
apiclient==1.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
distlib==0.3.6
filelock==3.8.0
Flask-SSLify==0.1.5
fonttools==4.38.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
magic-filter==1.0.9
MarkupSafe==2.1.1
matplotlib==3.6.3
multidict==6.0.4
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
platformdirs==2.5.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyngrok==5.2.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
PyTurboJPEG==1.6.7
pytz==2022.7
pyusb==1.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
